refactor(rag): replace HasSystem with strict AllSystemOnly semantics
对齐 Plan v2 的 isSystemOnly 严格守卫：仅当所有子问题都纯命中 SYSTEM 候选时才置位，
混合 SYSTEM+KB 场景继续走 RAG 检索。同时为 channel/intent_resolver 补充包级文档与
方法注释，便于后续维护。retriever.go 顺手修一处冗余的 float 转换。

+++ internal/rag/retriever.go
@@ -72,7 +72,7 @@ func parseSearchResults(results []milvusclient.SearchResult, collectionName stri
 
 	for i := 0; i < result.ResultCount; i++ {
 		chunk := RetrievedChunk{
-			Score:          float64(result.Scores[i]),
+			Score:          result.Scores[i],
 			CollectionName: collectionName,
 		}
 		if idCol != nil {
